WebBASIC 1.0
This is WebBASIC, which is a new scripting language.  However, you have
to download webbasic.js and put it on your website to use it in your
website
To set it up, type this in your page's HTML:
<script
src="replace_with_the_folder_that_contains_your_webbasic_script/webbasic.js"></script>
<script type="webbasic>

</script>
Note:Your webbasic scripts have to be in the first 10 scripts in your
page to work.
Commands:
ECHO string
Prints string.  No quotes needed.
GOTO n
go to line n
IF a case b THEN command(not IF)
If a case b is true, do command.
Cases:
*EQUAL: equal to
*UNEQUAL: unequal to
*LESS: less than
*GREATER: greater than
*GOE: greater than or equal to
*LOE: less than or equal to
Prints are in console.